re-arrange code
